blog_agro/requirements.txt
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
config==0.5.0.post0
coreapi==2.3.3
coreschema==0.0.4
decouple==0.0.7
dj-database-url==0.5.0
Django==3.2.4
django-cors-headers==3.7.0
django-filter==2.4.0
django-on-heroku==1.1.2
djangorestframework==3.12.4
drf-yasg==1.20.0
gunicorn==20.1.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
packaging==20.9
Pillow==8.2.0
psycopg2-binary==2.8.6
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
ruamel.yaml==0.17.7
ruamel.yaml.clib==0.2.2
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.5
v==0.0.0
whitenoise==5.2.0
